refactor: move current `createMiddleware` implementation to `src/middleware-legacy`
