+++ test/unit/middleware-constructor-test.ts
@@ -1,4 +1,4 @@
-import { createMiddleware as Middleware } from "../../src/middleware";
+import { createMiddleware as Middleware } from "../../src/middleware-legacy";
 
 test("options: none", () => {
   expect(() => Middleware({})).toThrow();
